Add tests for JobService

+++ i1-quartz-monitor/src/main/java/com/thoughtworks/i1/quartz/domain/TriggerVO.java
@@ -129,7 +129,7 @@ public class TriggerVO {
             return this;
         }
 
-        public TriggerVOBuilder repeat(int interval, int repeatCount) {
+        public TriggerVOBuilder repeat(long interval, int repeatCount) {
             this.repeatInterval = interval;
             this.repeatCount = repeatCount;
             return this;
